Split out dynlib loading from load-package
